Ajout le nouveau layout CollectivitePageLayout aux plans d'action
