+++ app.territoiresentransitions.react/src/app/pages/collectivite/PlansActions/PlansActions.tsx
@@ -1,23 +1,62 @@
 import {useCurrentCollectivite} from 'core-logic/hooks/useCurrentCollectivite';
-import PlansActionsNavigation from './PlansActionsNavigation';
 import {PlansActionsRoutes} from './PlansActionsRoutes';
+import CollectivitePageLayout from '../CollectivitePageLayout/CollectivitePageLayout';
+import {
+  generatePlanActionNavigationLinks,
+  usePlansNavigation,
+} from './PlanAction/data/usePlansNavigation';
+import {useFichesNonClasseesListe} from './FicheAction/data/useFichesNonClasseesListe';
+import {useCreateFicheAction} from './FicheAction/data/useUpsertFicheAction';
+import {useCreatePlanAction} from './PlanAction/data/useUpsertAxe';
 
 const PlansActions = () => {
   const collectivite = useCurrentCollectivite();
 
+  const {data: planListe} = usePlansNavigation();
+  const fichesNonClasseesListe = useFichesNonClasseesListe(
+    collectivite?.collectivite_id!
+  );
+
+  const {mutate: createFicheAction} = useCreateFicheAction();
+
+  const {mutate: createPlanAction} = useCreatePlanAction();
+
+  if (!collectivite) return null;
+
   return (
-    <div data-test="PlansAction" className="fr-container !px-0">
-      <div className="flex items-start">
-        {collectivite && (
+    <CollectivitePageLayout
+      sideNav={{
+        links: generatePlanActionNavigationLinks(
+          collectivite.collectivite_id,
+          planListe,
+          fichesNonClasseesListe?.count
+        ),
+        actions: !collectivite.readonly && (
           <>
-            <PlansActionsNavigation collectivite={collectivite} />
-            <PlansActionsRoutes
-              collectivite_id={collectivite.collectivite_id}
-            />
+            <li className="fr-sidemenu_item p-0 list-none">
+              <button
+                data-test="CreerFicheAction"
+                className="fr-btn fr-btn--secondary"
+                onClick={() => createFicheAction()}
+              >
+                Créer une fiche action
+              </button>
+            </li>
+            <li className="fr-sidemenu_item mt-6 p-0 list-none">
+              <button
+                data-test="AjouterPlanAction"
+                className="fr-btn fr-btn--tertiary"
+                onClick={() => createPlanAction()}
+              >
+                Ajouter un plan d'action
+              </button>
+            </li>
           </>
-        )}
-      </div>
-    </div>
+        ),
+      }}
+    >
+      <PlansActionsRoutes collectivite_id={collectivite.collectivite_id} />
+    </CollectivitePageLayout>
   );
 };
